backend/requirements.txt
fastapi
uvicorn
groq
chromadb
voyageai
pymupdf
python-docx
python-multipart
langchain-text-splitters